Introducing per-partition PartitionStore
**IMPORTANT:** This breaks queries through datafusion until we workout how data fusion will shard queries across partitions.
