+++ crates/storage-rocksdb/Cargo.toml
@@ -25,9 +25,10 @@ codederror = { workspace = true }
 derive_builder = { workspace = true }
 derive_more = { workspace = true }
 drain = { workspace = true }
+enum-map = { workspace = true }
 futures = { workspace = true }
 futures-util = { workspace = true }
-once_cell = "1.18.0"
+once_cell = { workspace = true }
 paste = { workspace = true }
 prost = { workspace = true }
 rocksdb = { workspace = true }
@@ -40,6 +41,7 @@ sync_wrapper = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
 restate-core = { workspace = true, features = ["test-util"] }
