+++ crates/storage-rocksdb/benches/basic_benchmark.rs
@@ -8,6 +8,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use std::ops::RangeInclusive;
+
 use criterion::{criterion_group, criterion_main, Criterion};
 use restate_core::TaskCenterBuilder;
 use restate_rocksdb::RocksDbManager;
@@ -15,22 +17,13 @@ use restate_storage_api::deduplication_table::{
     DedupSequenceNumber, DeduplicationTable, ProducerId,
 };
 use restate_storage_api::Transaction;
-use restate_storage_rocksdb::RocksDBStorage;
+use restate_storage_rocksdb::{OpenMode, PartitionStore, PartitionStoreManager};
 use restate_types::arc_util::Constant;
 use restate_types::config::{CommonOptions, WorkerOptions};
+use restate_types::identifiers::PartitionKey;
 use tokio::runtime::Builder;
 
-async fn writing_to_rocksdb(worker_options: WorkerOptions) {
-    //
-    // setup
-    //
-    let mut rocksdb = RocksDBStorage::open(
-        Constant::new(worker_options.storage.clone()),
-        Constant::new(worker_options.storage.rocksdb),
-    )
-    .await
-    .expect("RocksDB storage creation should succeed");
-
+async fn writing_to_rocksdb(mut rocksdb: PartitionStore) {
     //
     // write
     //
@@ -52,16 +45,38 @@ fn basic_writing_reading_benchmark(c: &mut Criterion) {
         .build()
         .expect("task_center builds");
 
+    let worker_options = WorkerOptions::default();
     tc.run_in_scope_sync("db-manager-init", None, || {
         RocksDbManager::init(Constant::new(CommonOptions::default()))
     });
+    let rocksdb = tc.block_on("test-setup", None, async {
+        //
+        // setup
+        //
+        let manager = PartitionStoreManager::create(
+            Constant::new(worker_options.storage.clone()),
+            Constant::new(worker_options.storage.rocksdb.clone()),
+            &[(0, RangeInclusive::new(0, PartitionKey::MAX))],
+        )
+        .await
+        .expect("DB creation succeeds");
+        manager
+            .open_partition_store(
+                0,
+                RangeInclusive::new(0, PartitionKey::MAX),
+                OpenMode::CreateIfMissing,
+                &worker_options.storage.rocksdb,
+            )
+            .await
+            .expect("column family is open")
+    });
+
     let mut group = c.benchmark_group("RocksDB");
     group.sample_size(10).bench_function("writing", |bencher| {
         // This will generate a temp dir since we have test-util feature enabled
-        let worker_options = WorkerOptions::default();
         bencher
             .to_async(&rt)
-            .iter(|| writing_to_rocksdb(worker_options.clone()));
+            .iter(|| writing_to_rocksdb(rocksdb.clone()));
     });
 
     group.finish();
